+++ src/main/CollisionChecker.java
@@ -12,10 +12,10 @@ public class CollisionChecker {
 
     public void checkTile(Entity entity) {
         
-        int entityLeftWorldX = entity.worldX + entity.solidArea.x;
-        int entityRightWorldX = entity.worldX + entity.solidArea.x + entity.solidArea.width;
-        int entityTopWorldY = entity.worldY + entity.solidArea.y;
-        int entityBottomWorldY = entity.worldY + entity.solidArea.y + entity.solidArea.height;
+        int entityLeftWorldX = (int) entity.worldX + entity.solidArea.x;
+        int entityRightWorldX = (int) entity.worldX + entity.solidArea.x + entity.solidArea.width;
+        int entityTopWorldY = (int) entity.worldY + entity.solidArea.y;
+        int entityBottomWorldY = (int) entity.worldY + entity.solidArea.y + entity.solidArea.height;
 
         int entityLeftCol = entityLeftWorldX/gp.tileSize;
         int entityRightCol = entityRightWorldX/gp.tileSize;
