+++ src/object/SuperObject.java
@@ -22,8 +22,8 @@ public class SuperObject {
     public void draw(Graphics2D g2, GamePanel gp) {
 
         //Player is always at the center of the screen, objects are drawn to the screen relative to the player world position
-        int screenX = worldX - gp.player.worldX + gp.player.screenX;
-        int screenY = worldY - gp.player.worldY + gp.player.screenY;
+        int screenX = (int) (worldX - gp.player.worldX + gp.player.screenX);
+        int screenY = (int) (worldY - gp.player.worldY + gp.player.screenY);
 
         //Check if the object would appear on screen before rendering
         if (worldX + gp.tileSize > gp.player.worldX - gp.player.screenX &&
